+++ WorkloadManagementSystem/Service/PilotsLoggingHandler.py
@@ -14,9 +14,11 @@ __RCSID__ = "$Id: $"
 
 from DIRAC.WorkloadManagementSystem.DB.PilotsLoggingDB import PilotsLoggingDB
 
+import types
 from DIRAC.Core.DISET.RequestHandler import RequestHandler
 from DIRAC import gLogger, S_OK, S_ERROR
 from DIRAC.Core.Utilities import Time
+# from DIRAC.WorkloadManagementSystem.DB import PilotAgentsDB
 
 class PilotsLoggingHandler( RequestHandler ):
 
@@ -27,17 +29,20 @@ class PilotsLoggingHandler( RequestHandler ):
   def initialize(self):
     self.pilotsLogging = PilotsLoggingDB()
 
-  types_addPilotsLogging = [ basestring, basestring, basestring, float, basestring ]
+  auth_addPilotsLogging = [ 'all' ]
+  types_addPilotsLogging = [ types.StringType, types.StringType, types.StringType, types.FloatType, types.StringType ]
   def export_addPilotsLogging( self, pilotUUID, status, minorStatus, timeStamp, source ):
 
     return self.pilotsLogging.addPilotsLogging( pilotUUID, status, minorStatus, timeStamp, source )
 
-  types_getPilotsLogging = [ [int, long] ]
+  auth_getPilotsLogging = [ 'all' ]
+  types_getPilotsLogging = [ types.IntType ]
   def export_getPilotsLogging( self, pilotID ):
 
     return self.pilotsLogging.getPilotsLogging( pilotID )
 
-  types_setPilotsUUIDtoIDMapping = [ basestring, [int, long] ]
+  auth_setPilotsUUIDtoIDMapping = [ 'all' ]
+  types_setPilotsUUIDtoIDMapping = [ types.StringType, types.IntType ]
   def export_setPilotsUUIDtoIDMapping( self, pilotUUID, pilotID ):
 
     return self.pilotsLogging.setPilotsUUIDtoIDMapping( pilotUUID, pilotID )
@@ -48,7 +53,8 @@ class PilotsLoggingHandler( RequestHandler ):
 
     return self.pilotsLogging.addPilotsUUID( pilotUUID )
 
-  types_deletePilotsLogging = [ [int, long,  list] ]
-  def export_deletePilotsLogging( self, pilotID ):
+  auth_detelePilotsLogging = [ 'all' ]
+  types_detelePilotsLogging = [ types.IntType ]
+  def export_detelePilotsLogging( self, pilotID ):
 
     return self.pilotsLogging.deletePilotsLogging( pilotID )
